# openapi.json
{"openapi":"3.0.1","info":{"title":"Portalku API","description":"API portal informasi akademik - latihan MK Pemrograman Web Lanjut, D3 Teknologi Informasi PSDKU Lumajang","contact":{"name":"D3 TI PSDKU Lumajang"},"version":"v1"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"paths":{"/api/pengumuman/{id}":{"get":{"tags":["pengumuman-controller"],"summary":"Menampilkan satu pengumuman","description":"Mengambil data pengumuman berdasarkan ID.","operationId":"satu","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"Data pengumuman berhasil ditemukan","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PengumumanResponse"}}}}}},"put":{"tags":["pengumuman-controller"],"summary":"Mengubah pengumuman","description":"Mengubah data pengumuman berdasarkan ID.","operationId":"ubah","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PengumumanRequest"}}},"required":true},"responses":{"200":{"description":"Data pengumuman berhasil diubah","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PengumumanResponse"}}}}}},"delete":{"tags":["pengumuman-controller"],"summary":"Menghapus pengumuman","description":"Menghapus data pengumuman berdasarkan ID.","operationId":"hapus","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"204":{"description":"Data pengumuman berhasil dihapus"}}}},"/api/pengumuman":{"get":{"tags":["pengumuman-controller"],"summary":"Menampilkan semua pengumuman","description":"Mengambil seluruh data pengumuman yang tersedia.","operationId":"semua","responses":{"200":{"description":"Data pengumuman berhasil diambil","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/PengumumanRingkasResponse"}}}}}}},"post":{"tags":["pengumuman-controller"],"summary":"Menambahkan pengumuman","description":"Menambahkan data pengumuman baru ke dalam sistem.","operationId":"tambah","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PengumumanRequest"}}},"required":true},"responses":{"201":{"description":"Data pengumuman berhasil ditambahkan","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PengumumanResponse"}}}}}}},"/api/jurnal":{"get":{"tags":["jurnal-controller"],"summary":"Menampilkan semua jurnal","description":"Mengambil seluruh data jurnal yang tersedia.","operationId":"semua_1","responses":{"200":{"description":"Data jurnal berhasil diambil","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Jurnal"}}}}}}},"post":{"tags":["jurnal-controller"],"summary":"Menambahkan jurnal","description":"Menambahkan data jurnal baru ke dalam sistem.","operationId":"tambah_1","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Jurnal"}}},"required":true},"responses":{"200":{"description":"Data jurnal berhasil ditambahkan","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Jurnal"}}}}}}},"/api/prodi/uji":{"get":{"tags":["prodi-controller"],"operationId":"uji","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ProdiRingkasDto"}}}}}}},"/api/kurikulum/uji":{"get":{"tags":["kurikulum-controller"],"operationId":"uji_1","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Kurikulum"}}}}}}}},"/api/jurnal/{id}":{"get":{"tags":["jurnal-controller"],"summary":"Menampilkan satu jurnal","description":"Mengambil data jurnal berdasarkan ID.","operationId":"satu_1","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"Data jurnal berhasil ditemukan","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Jurnal"}}}}}},"delete":{"tags":["jurnal-controller"],"summary":"Menghapus jurnal","description":"Menghapus data jurnal berdasarkan ID.","operationId":"hapus_1","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"204":{"description":"Data jurnal berhasil dihapus"}}}},"/api/info":{"get":{"tags":["info-controller"],"operationId":"info","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"object"}}}}}}}},"/api/halo":{"get":{"tags":["halo-controller"],"summary":"Cek status layanan Portalku","description":"Mengembalikan pesan bahwa layanan Portalku sedang berjalan.","operationId":"halo","responses":{"200":{"description":"Layanan berjalan dengan baik","content":{"*/*":{"schema":{"type":"string"}}}}}}}},"components":{"schemas":{"PengumumanRequest":{"required":["idKategori","isi","judul","tanggalTerbit"],"type":"object","properties":{"judul":{"type":"string"},"isi":{"type":"string"},"tanggalTerbit":{"type":"string","format":"date"},"idKategori":{"type":"integer","format":"int32"}}},"PengumumanResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"judul":{"type":"string"},"isi":{"type":"string"},"tanggalTerbit":{"type":"string","format":"date"},"jumlahDilihat":{"type":"integer","format":"int32"},"namaKategori":{"type":"string"},"jumlahLampiran":{"type":"integer","format":"int32"}}},"Jurnal":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"judul":{"type":"string"},"penerbit":{"type":"string"},"tahunTerbit":{"type":"integer","format":"int32"},"issn":{"type":"string"}}},"ProdiRingkasDto":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"kode":{"type":"string"},"nama":{"type":"string"},"jenjang":{"type":"string","enum":["D3","D4"]},"jumlahKurikulum":{"type":"integer","format":"int32"}}},"PengumumanRingkasResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"judul":{"type":"string"},"namaKategori":{"type":"string"},"tanggalTerbit":{"type":"string","format":"date"},"jumlahDilihat":{"type":"integer","format":"int32"}}},"Kategori":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"nama":{"type":"string"},"daftarPengumuman":{"type":"array","items":{"$ref":"#/components/schemas/Pengumuman"}}}},"Kurikulum":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"kode":{"type":"string"},"tahunBerlaku":{"type":"integer","format":"int32"},"status":{"type":"string","enum":["AKTIF","NONAKTIF"]},"prodi":{"$ref":"#/components/schemas/Prodi"}}},"Lampiran":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"namaBerkas":{"type":"string"},"ukuran":{"type":"integer","format":"int64"},"pengumuman":{"$ref":"#/components/schemas/Pengumuman"}}},"Pengumuman":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"judul":{"type":"string"},"isi":{"type":"string"},"tanggalTerbit":{"type":"string","format":"date"},"jumlahDilihat":{"type":"integer","format":"int32"},"kategori":{"$ref":"#/components/schemas/Kategori"},"daftarLampiran":{"type":"array","items":{"$ref":"#/components/schemas/Lampiran"}},"targetProdi":{"uniqueItems":true,"type":"array","items":{"$ref":"#/components/schemas/Prodi"}}}},"Prodi":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"kode":{"type":"string"},"nama":{"type":"string"},"jenjang":{"type":"string","enum":["D3","D4"]},"daftarKurikulum":{"type":"array","items":{"$ref":"#/components/schemas/Kurikulum"}},"daftarPengumuman":{"uniqueItems":true,"type":"array","items":{"$ref":"#/components/schemas/Pengumuman"}}}}}}}